UPBGE: Use std::vector for vertex and texture attributs instead of static array stored in RAS_OpenGLRasterizer.
Previously two static arrays was used to store texture and vertex attributs in RAS_OpenGLRasterizer mutated by
SetTexCoordNum, SetAttribNum, SetTexCoord, SetAttrib.

We now replace this system by the usage of std::vector of RAS_IRasterizer::TexCoGen/Attrib acceded by
RAS_IPolyMaterial::GetTexCoGenList/GetAttribList. The advantage is to don't generate the attributs every frame but
store them in a list during the material conversion. To follow the existing behaviour KX_BlenderMaterial store two
list pointer cooresponding to the current texture and vertex attributs used for the current frame, ths choose is
done in KX_BlenderMaterial::Activate.
Vertex attributs list are stored in both BL_Shader (if use tangent attribut) and BL_BlenderShader and get with the
same named function : GetAttribList.
For texture attributs (used in multitexture mode), we treat them only in KX_BlenderMaterial, they are generated at
instancing and stored in m_texCoGenList and its pointer is set to m_currentTexCoGenList during material activation
if needed.

This new system is easily to maintain but a bit slower or faster in certain cases. The tests was done with 10000
objects of 42 vertexes, 80 triangles and of four materials. If the 4 materials are without alpha so without material
switchs the result is 168 versus 165 (+1.8%), whereas if all materials use transparency with alpha blend and mesh
ordering with a lot of material switchs the result are 1000 versus 1066 (-6.6%).

The speedup for alpha material and the code simplification is enougth to accept the patch and the little slowdown.